feat: Added refreshToken route
